# Conflicts:
#	3DFProject_Official/Assets/AmplifyShaderEditor/Examples/Built-In Samples.unitypackage.meta
#	3DFProject_Official/Assets/MaterialMikeRequire/Materials/MI_Wall_Brick.mat
#	3DFProject_Official/Assets/Scenes/Test Arno2.0.unitypackage.meta
#	3DFProject_Official/Assets/Scenes/Test_Rick.unity.meta
#	3DFProject_Official/UserSettings/EditorUserSettings.asset

+++ 3DFProject_Official/Assets/Scripts/Player/MouseLook.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class MouseLook : Singleton<MouseLook>
 {
-    public float mouseSensitivity = 100f;
+    public int mouseSensitivity = 100;
 
     [SerializeField]private Transform playerBody;
     [SerializeField]private CinemachineVirtualCamera[] cams;
